Revised definition of and examples to the periodic convolution, minor edits, common plotting style, re-run examples, included FFT/fast convolution notebooks in CI

+++ discrete_fourier_transform/tools.py
@@ -13,10 +13,10 @@ def cconv(x, y, P):
 def _wrap(x, N):
     """Zero-padding to length N or periodic summation with period N"""
     M = len(x)
-    rows = np.ceil(M/N)
+    rows = np.int(np.ceil(M/N))
     
     if (M < int(N*rows)):
-        x = np.pad(x, (0, int(N*rows-M)), 'constant')
+        x = np.pad(x, (0, N*rows-M), 'constant')
     
     x = np.reshape(x, (rows, N))
